refactor(protocol): split work item contracts
