+++ docs/implementation-status.md
@@ -41,11 +41,11 @@ Single source of truth for the gap between accepted design and running code. Oth
 | ReadBackCheck evidence records + store helper | ✅ | `packages/protocol/src/work-item/`, `packages/session/src/work-item/` | Shipped structured `ReadBackCheck` records for URL fetch, API query, and citation match observations plus `WorkItemStore.addReadBackEvidence()` to persist read-back results as verification evidence |
 | Runtime read-back executors (re-fetch URL, re-query API, citation-in-source matching) | ✅ | `packages/openomni/src/evidence/` | Shipped explicit `ReadBackExecutor.execute()` / `record()` API for HTTP URL re-fetch, read-only HTTP API re-query, and citation-in-source matching; `record()` feeds runtime observations into `WorkItemStore.addReadBackEvidence()` |
 | Automatic read-back invocation during completion gating | ✅ | `packages/openomni/src/dispatch/handlers/worker-completion.ts` | `worker.spawn` succeeded results may include `readBackRequests`; dispatch now runs each request through `ReadBackExecutor.record()` before completion, persists verification evidence, attaches generated evidence IDs to completion-report claims, and blocks completion if read-back evidence fails |
-| Connector/channel-specific read-back request builders | 🚧 | `packages/protocol/src/work-item/index.ts`, `packages/protocol/src/app-connector/index.ts`, `apps/server/src/connector/read-back-builder.ts`, `apps/server/src/connector/process-driver.ts` | AppConnector completion metadata can now declare templated read-back request builders, and the connector process driver fail-closes missing/non-JSON builder sources while appending generated `readBackRequests` to completion envelopes before the existing completion gate runs. Channel adapter builders for outbound publish/send/API actions remain pending |
+| Connector/channel-specific read-back request builders | 🚧 | `packages/protocol/src/work-item/`, `packages/protocol/src/app-connector/index.ts`, `apps/server/src/connector/read-back-builder.ts`, `apps/server/src/connector/process-driver.ts` | AppConnector completion metadata can now declare templated read-back request builders, and the connector process driver fail-closes missing/non-JSON builder sources while appending generated `readBackRequests` to completion envelopes before the existing completion gate runs. Channel adapter builders for outbound publish/send/API actions remain pending |
 | Internal worker retry default + kernel-enforced exhaustion blocker | ✅ | `packages/session/src/work-item/`, `packages/openomni/src/dispatch/handlers/worker.ts` | Internal `worker.spawn` ledger items are tagged `executorKind: internal_chat_agent`, default to `maxAttempts: 3`, and `WorkItemStore.retry()` refuses exhausted items after recording a `waiting_input` escalation blocker. Owner-visible notification / failed-item surfacing remains pending below |
 | Owner-visible retry exhaustion ledger surface | ✅ | `apps/server/src/handler/conversation.ts` | Authenticated local WebSocket `show open tasks` includes failed WorkItems with an active retry-exhaustion Owner escalation blocker, including attempt/max-attempt context, so exhausted internal worker retries are no longer invisible |
 | App manifest retry defaults + human reminder policy + non-local Owner notification | 📋 | — | ADR-011 — CLI apps should inherit per-application manifest defaults; humans are not retried, they need reminder policy under the social budget; exhausted failed items still need push-style Owner notification beyond the local ledger surface |
-| Owner adoption signal storage primitive (`outcome`: adopted/corrected/redone/ignored) | ✅ | `packages/protocol/src/work-item/index.ts`, `packages/session/src/work-item/outcome.ts` | `WorkItem.Info.outcome` is schema-backed and `WorkItemStore.recordOutcome()` records Owner feedback on completed WorkItems while publishing `work_item.outcome_recorded` for downstream Governor consumers |
+| Owner adoption signal storage primitive (`outcome`: adopted/corrected/redone/ignored) | ✅ | `packages/protocol/src/work-item/`, `packages/session/src/work-item/outcome.ts` | `WorkItem.Info.outcome` is schema-backed and `WorkItemStore.recordOutcome()` records Owner feedback on completed WorkItems while publishing `work_item.outcome_recorded` for downstream Governor consumers |
 | Owner adoption collection surfaces + Governor calibration | 📋 | — | ADR-011 — no Owner-facing adopted/corrected/redone/ignored capture UI or dispatch command yet; Governor does not aggregate adoption outcomes or calibrate Resident evaluation leniency |
 | Task ledger view ("show open tasks" — the OS's `ps`) | ✅ | `apps/server/src/handler/conversation.ts` | Authenticated local WebSocket chat command returns a capped pending/running/blocked WorkItem diagnostic plus retry-exhausted failed-item escalations from `WorkItemStore`; unauthenticated WebSocket and external-channel actor-scoped visibility remain future work because current `WorkItem` / `InboundMessage` schemas have no owner, trust-tier, or channel-grant field; web view also remains future work |
 | Bus persistence (hash-chained event journal) | ✅ | `packages/session/src/bus-persistence/` | — |
