+++ packages/protocol/AGENTS.md
@@ -44,7 +44,7 @@ src/
 - **Policy timings**: 14 policy timing points — `inbound.receive`, `run.start`, `turn.start`, `context.prepare`, `resources.prepare`, `model.request`, `model.response`, `invoke.prepare`, `invoke.result`, `turn.finish`, `completion.prepare`, `writeback.commit`, `run.finish`, `error`. `Policy.PolicyDecision` verdict is one of `allow | deny | pending`; legacy permission evaluators still use `EvaluationResult.action` (`continue | abort`).
 - **Subagent lifecycle**: `Subagent.Events.*` covers worker sessions (`WorkerSessionSpawned/Resumed/Cancelled`), worker runs (`WorkerRunStarted/Completed/Failed`), consultations (`WorkerConsultationRequested/Completed`), and background tasks (`BackgroundTaskLaunched/Completed/Failed/Cancelled`).
 - **Storage sub-adapters**: `Storage.WorkItemSubAdapter` in `storage/index.ts` — pure interface contract with no runtime logic. Implementation lives in `@openomni/session`.
-- **WorkItem namespace**: `work-item/index.ts` defines `WorkItem.Info` (universal work state schema with derived status, routing/session fields, completion reports, retry/outcome fields, blockers, evidence, read-back checks, verification gates), `WorkItem.ReadBackCheck` (URL fetch / API query / citation match observations), `WorkItem.Events.*` (Created, Updated, StatusChanged, Completed, Failed, Removed via `BusEvent.define()`), `deriveStatus()` (pure status derivation from timestamps/blockers), and `generateHash()` (base36 12-char collision-safe IDs).
+- **WorkItem namespace**: `work-item/index.ts` is the public facade. `work-item/schemas.ts` defines `WorkItem.Info` (universal work state schema with derived status, routing/session fields, completion reports, retry/outcome fields, blockers, evidence, read-back checks, verification gates), `WorkItem.ReadBackCheck` (URL fetch / API query / citation match observations), and related schemas. `work-item/events.ts` defines `WorkItem.Events.*` (Created, Updated, StatusChanged, Completed, Failed, Removed via `BusEvent.define()`), `work-item/status.ts` defines `deriveStatus()` (pure status derivation from timestamps/blockers), and `work-item/hash.ts` defines `generateHash()` (base36 12-char collision-safe IDs).
 - **Execution/IPC contracts**: `execution/`, `ipc/`, and `worker-bootstrap/` describe worker requests, responses, and bootstrap payloads only. Runtime worker lifecycle lives in `@openomni/coordinator`.
 - **AppConnector namespace**: `app-connector/index.ts` defines installed-app connector schema contracts. Runtime install, consent, and process execution live above protocol.
 - **Trace contract**: `trace/index.ts` defines the shared shape; helper creation lives in `@openomni/session`.
@@ -76,8 +76,8 @@ Keep these as protocol contracts only. Runtime policy and storage implementation
 - Adding a new policy timing? Update `Policy.Timing` in `policy/index.ts` and coordinate with `packages/agent/src/core/policy/engine.ts`.
 - Adding a new subagent event? Extend `Subagent.Events` in `subagent/index.ts` with a `BusEvent.define()` call.
 - Adding a new storage sub-adapter interface? Add it to `storage/index.ts` as a named interface under the `Storage` namespace.
-- Adding a work-item field? Update `WorkItem.Info` in `work-item/index.ts`. If it affects status derivation, update `deriveStatus()`.
-- Adding a work-item event? Extend `WorkItem.Events` in `work-item/index.ts` with a `BusEvent.define()` call.
+- Adding a work-item field? Update `WorkItem.Info` in `work-item/schemas.ts`. If it affects status derivation, update `deriveStatus()` in `work-item/status.ts`.
+- Adding a work-item event? Extend `WorkItem.Events` in `work-item/events.ts` with a `BusEvent.define()` call.
 - Adding a new worker request or IPC field? Update `execution/`, `ipc/`, or `worker-bootstrap/` here first, then adapt coordinator/openomni/server callers.
 - Adding trace metadata? Update `trace/index.ts`; helper functions stay in `@openomni/session`.
 - This package builds to `dist/` — run `bun run build` after changes.
